Download button

Wrapper around `bsplus::bs_button()` to provide a download button for HTML outputs in R Markdown. Internally, the function writes the file to `tempdir()`, encodes it, and produces the download button. Currently, Internet Explorer does not support downloading embedded files.

download_this(
  .data,
  output_name,
  output_extension = c(".csv", ".xlsx", ".rds"),
  button_label = "Download data",
  button_type = c("default", "primary", "success", "info", "warning", "danger"),
  has_icon = TRUE,
  icon = "fa fa-save",
  self_contained = FALSE,
  ...
)

Arguments

.data
A data frame or (named) list to write to disk. See 'Examples' for more details.

output_name
Name of of the output file.

output_extension
Extension of the output file. Currently, `.csv`, `.xlsx`, and `.rds` are supported. If a (named) list is passed to the function, only `.xlsx` and `.rds` are supported.

button_label
Character (HTML), button label

button_type
Character, one of the standard Bootstrap types

has_icon
Specify whether to include fontawesome icons in the button label

icon
Fontawesome tag e.g.: "fa fa-save"

self_contained
A boolean to specify whether your HTML output is self-contained. Default to `FALSE`.

...
attributes (named arguments) and children (unnamed arguments) of the button, passed to `htmltools::tag()`.

Value

htmltools::tag, <button>

Warning

This example will write the `mtcars` dataset to `tempdir()` and produce the download button for the file `mtcars dataset.csv` with the `fa fa-save` icon on the `Download data` label.

Examples

# Passing a data frame to the function
mtcars %>%
  download_this(
    output_name = "mtcars dataset",
    output_extension = ".csv",
    button_label = "Download data",
    button_type = "warning",
    has_icon = TRUE,
    icon = "fa fa-save"
  )

#> <a href="data:text/csv;base64,bXBnO2N5bDtkaXNwO2hwO2RyYXQ7d3Q7cXNlYzt2czthbTtnZWFyO2NhcmIKMjEsMDs2OzE2MCwwOzExMDszLDkwOzIsNjIwOzE2LDQ2OzA7MTs0OzQKMjEsMDs2OzE2MCwwOzExMDszLDkwOzIsODc1OzE3LDAyOzA7MTs0OzQKMjIsODs0OzEwOCwwOyA5MzszLDg1OzIsMzIwOzE4LDYxOzE7MTs0OzEKMjEsNDs2OzI1OCwwOzExMDszLDA4OzMsMjE1OzE5LDQ0OzE7MDszOzEKMTgsNzs4OzM2MCwwOzE3NTszLDE1OzMsNDQwOzE3LDAyOzA7MDszOzIKMTgsMTs2OzIyNSwwOzEwNTsyLDc2OzMsNDYwOzIwLDIyOzE7MDszOzEKMTQsMzs4OzM2MCwwOzI0NTszLDIxOzMsNTcwOzE1LDg0OzA7MDszOzQKMjQsNDs0OzE0Niw3OyA2MjszLDY5OzMsMTkwOzIwLDAwOzE7MDs0OzIKMjIsODs0OzE0MCw4OyA5NTszLDkyOzMsMTUwOzIyLDkwOzE7MDs0OzIKMTksMjs2OzE2Nyw2OzEyMzszLDkyOzMsNDQwOzE4LDMwOzE7MDs0OzQKMTcsODs2OzE2Nyw2OzEyMzszLDkyOzMsNDQwOzE4LDkwOzE7MDs0OzQKMTYsNDs4OzI3NSw4OzE4MDszLDA3OzQsMDcwOzE3LDQwOzA7MDszOzMKMTcsMzs4OzI3NSw4OzE4MDszLDA3OzMsNzMwOzE3LDYwOzA7MDszOzMKMTUsMjs4OzI3NSw4OzE4MDszLDA3OzMsNzgwOzE4LDAwOzA7MDszOzMKMTAsNDs4OzQ3MiwwOzIwNTsyLDkzOzUsMjUwOzE3LDk4OzA7MDszOzQKMTAsNDs4OzQ2MCwwOzIxNTszLDAwOzUsNDI0OzE3LDgyOzA7MDszOzQKMTQsNzs4OzQ0MCwwOzIzMDszLDIzOzUsMzQ1OzE3LDQyOzA7MDszOzQKMzIsNDs0OyA3OCw3OyA2Njs0LDA4OzIsMjAwOzE5LDQ3OzE7MTs0OzEKMzAsNDs0OyA3NSw3OyA1Mjs0LDkzOzEsNjE1OzE4LDUyOzE7MTs0OzIKMzMsOTs0OyA3MSwxOyA2NTs0LDIyOzEsODM1OzE5LDkwOzE7MTs0OzEKMjEsNTs0OzEyMCwxOyA5NzszLDcwOzIsNDY1OzIwLDAxOzE7MDszOzEKMTUsNTs4OzMxOCwwOzE1MDsyLDc2OzMsNTIwOzE2LDg3OzA7MDszOzIKMTUsMjs4OzMwNCwwOzE1MDszLDE1OzMsNDM1OzE3LDMwOzA7MDszOzIKMTMsMzs4OzM1MCwwOzI0NTszLDczOzMsODQwOzE1LDQxOzA7MDszOzQKMTksMjs4OzQwMCwwOzE3NTszLDA4OzMsODQ1OzE3LDA1OzA7MDszOzIKMjcsMzs0OyA3OSwwOyA2Njs0LDA4OzEsOTM1OzE4LDkwOzE7MTs0OzEKMjYsMDs0OzEyMCwzOyA5MTs0LDQzOzIsMTQwOzE2LDcwOzA7MTs1OzIKMzAsNDs0OyA5NSwxOzExMzszLDc3OzEsNTEzOzE2LDkwOzE7MTs1OzIKMTUsODs4OzM1MSwwOzI2NDs0LDIyOzMsMTcwOzE0LDUwOzA7MTs1OzQKMTksNzs2OzE0NSwwOzE3NTszLDYyOzIsNzcwOzE1LDUwOzA7MTs1OzYKMTUsMDs4OzMwMSwwOzMzNTszLDU0OzMsNTcwOzE0LDYwOzA7MTs1OzgKMjEsNDs0OzEyMSwwOzEwOTs0LDExOzIsNzgwOzE4LDYwOzE7MTs0OzIK" download="mtcars dataset.csv">
#>   <button class="btn btn-warning"><i class="fa fa-save"></i> Download data</button>
#> </a>

# Passing a list with data frames to the function
list(mtcars, iris) %>%
  download_this(
    output_name = "mtcars and iris datasets",
    output_extension = ".csv",
    button_label = "Download data",
    button_type = "warning",
    has_icon = TRUE,
    icon = "fa fa-save"
  )

#> Error: I am sorry, lists are not supported in '.csv'. Please, choose '.xlsx' instead.

# Passing a named list with data frames to the function
list('mtcars dataset' = mtcars, 'iris dataset' = iris) %>%
  download_this(
    output_name = "mtcars and iris datasets",
    output_extension = ".csv",
    button_label = "Download data",
    button_type = "warning",
    has_icon = TRUE,
    icon = "fa fa-save"
  )

#> Error: I am sorry, lists are not supported in '.csv'. Please, choose '.xlsx' instead.