Download button

Wrapper around bsplus::bs_button() to provide a download button for HTML outputs in R Markdown. Internally, the function writes the file to tempdir(), encodes it, and produces the download button. Currently, Internet Explorer does not support downloading embedded files.

download_this(
  .data,
  output_name,
  output_extension = c(".csv", ".xlsx"),
  button_label = "Download data",
  button_type = c("default", "primary", "success", "info", "warning", "danger"),
  has_icon = TRUE,
  icon = "fa fa-save",
  ...
)

Arguments

.data
A data frame or (named) list to write to disk. See 'Examples' for more details.

output_name
Name of of the output file.

output_extension
Extension of the output file. Currently, only .csv and .xlsx are supported. If a (named) list is passed to the function, only .xlsx is supported.

button_label
character (HTML), button label

button_type
character, one of the standard Bootstrap types

has_icon
specify whether to include fontawesome icons in the button label

icon
fontawesome tag e.g.: "fa fa-save"

...
attributes (named arguments) and children (unnamed arguments) of the button, passed to htmltools::tag

Value

htmltools::tag, <button/>

Warning

This example will write the mtcars dataset to tempdir() and produce the download button for the file `mtcars dataset.csv` with the `fa fa-save` icon on the `Download data` label.

Examples

# Passing a data frame to the function
mtcars %>%
  download_this(
    output_name = "mtcars dataset",
    output_extension = ".csv",
    button_label = "Download data",
    button_type = "warning",
    has_icon = TRUE,
    icon = "fa fa-save"
  )

#> <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css" integrity="sha256-h20CPZ0QyXlBuAw7A+KluUYx/3pK+c7lYEpqLTlxjYQ=" crossorigin="anonymous" />
#> <a href="data:text/csv;base64,bXBnO2N5bDtkaXNwO2hwO2RyYXQ7d3Q7cXNlYzt2czthbTtnZWFyO2NhcmIKMjEsMDs2OzE2MCwwOzExMDszLDkwOzIsNjIwOzE2LDQ2OzA7MTs0OzQKMjEsMDs2OzE2MCwwOzExMDszLDkwOzIsODc1OzE3LDAyOzA7MTs0OzQKMjIsODs0OzEwOCwwOyA5MzszLDg1OzIsMzIwOzE4LDYxOzE7MTs0OzEKMjEsNDs2OzI1OCwwOzExMDszLDA4OzMsMjE1OzE5LDQ0OzE7MDszOzEKMTgsNzs4OzM2MCwwOzE3NTszLDE1OzMsNDQwOzE3LDAyOzA7MDszOzIKMTgsMTs2OzIyNSwwOzEwNTsyLDc2OzMsNDYwOzIwLDIyOzE7MDszOzEKMTQsMzs4OzM2MCwwOzI0NTszLDIxOzMsNTcwOzE1LDg0OzA7MDszOzQKMjQsNDs0OzE0Niw3OyA2MjszLDY5OzMsMTkwOzIwLDAwOzE7MDs0OzIKMjIsODs0OzE0MCw4OyA5NTszLDkyOzMsMTUwOzIyLDkwOzE7MDs0OzIKMTksMjs2OzE2Nyw2OzEyMzszLDkyOzMsNDQwOzE4LDMwOzE7MDs0OzQKMTcsODs2OzE2Nyw2OzEyMzszLDkyOzMsNDQwOzE4LDkwOzE7MDs0OzQKMTYsNDs4OzI3NSw4OzE4MDszLDA3OzQsMDcwOzE3LDQwOzA7MDszOzMKMTcsMzs4OzI3NSw4OzE4MDszLDA3OzMsNzMwOzE3LDYwOzA7MDszOzMKMTUsMjs4OzI3NSw4OzE4MDszLDA3OzMsNzgwOzE4LDAwOzA7MDszOzMKMTAsNDs4OzQ3MiwwOzIwNTsyLDkzOzUsMjUwOzE3LDk4OzA7MDszOzQKMTAsNDs4OzQ2MCwwOzIxNTszLDAwOzUsNDI0OzE3LDgyOzA7MDszOzQKMTQsNzs4OzQ0MCwwOzIzMDszLDIzOzUsMzQ1OzE3LDQyOzA7MDszOzQKMzIsNDs0OyA3OCw3OyA2Njs0LDA4OzIsMjAwOzE5LDQ3OzE7MTs0OzEKMzAsNDs0OyA3NSw3OyA1Mjs0LDkzOzEsNjE1OzE4LDUyOzE7MTs0OzIKMzMsOTs0OyA3MSwxOyA2NTs0LDIyOzEsODM1OzE5LDkwOzE7MTs0OzEKMjEsNTs0OzEyMCwxOyA5NzszLDcwOzIsNDY1OzIwLDAxOzE7MDszOzEKMTUsNTs4OzMxOCwwOzE1MDsyLDc2OzMsNTIwOzE2LDg3OzA7MDszOzIKMTUsMjs4OzMwNCwwOzE1MDszLDE1OzMsNDM1OzE3LDMwOzA7MDszOzIKMTMsMzs4OzM1MCwwOzI0NTszLDczOzMsODQwOzE1LDQxOzA7MDszOzQKMTksMjs4OzQwMCwwOzE3NTszLDA4OzMsODQ1OzE3LDA1OzA7MDszOzIKMjcsMzs0OyA3OSwwOyA2Njs0LDA4OzEsOTM1OzE4LDkwOzE7MTs0OzEKMjYsMDs0OzEyMCwzOyA5MTs0LDQzOzIsMTQwOzE2LDcwOzA7MTs1OzIKMzAsNDs0OyA5NSwxOzExMzszLDc3OzEsNTEzOzE2LDkwOzE7MTs1OzIKMTUsODs4OzM1MSwwOzI2NDs0LDIyOzMsMTcwOzE0LDUwOzA7MTs1OzQKMTksNzs2OzE0NSwwOzE3NTszLDYyOzIsNzcwOzE1LDUwOzA7MTs1OzYKMTUsMDs4OzMwMSwwOzMzNTszLDU0OzMsNTcwOzE0LDYwOzA7MTs1OzgKMjEsNDs0OzEyMSwwOzEwOTs0LDExOzIsNzgwOzE4LDYwOzE7MTs0OzIK" download="mtcars dataset.csv">
#>   <button class="btn btn-warning"><i class="fa fa-save"></i> Download data</button>
#> </a>

# Passing a list with data frames to the function
list(mtcars, iris) %>%
  download_this(
    output_name = "mtcars and iris datasets",
    output_extension = ".csv",
    button_label = "Download data",
    button_type = "warning",
    has_icon = TRUE,
    icon = "fa fa-save"
  )

#> <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css" integrity="sha256-h20CPZ0QyXlBuAw7A+KluUYx/3pK+c7lYEpqLTlxjYQ=" crossorigin="anonymous" />
#> <a href="data:application/vnd.openxmlformats-officedocument.spreadsheetml.sheet;base64,UEsDBBQAAAAIAAAAIQCeDieWQgEAAI8EAAATAAAAW0NvbnRlbnRfVHlwZXNdLnhtbLWUy27CMBBF9/2KyNsqNrCoqorAoo9ly4J+gGtPiIVf8hgKf99JQpFaUSii3cSK5849dyzL4+nG2WINCU3wFRvyASvAq6CNX1Tsdf5U3rICs/Ra2uChYltANp1cjefbCFhQs8eKNTnHOyFQNeAk8hDBU6UOyclMv2kholRLuQAxGgxuhAo+g89lbj3YZPwAtVzZXDxuaLsPksAiK+57YcuqmIzRGiUz1cXa62+Uckfg1NlpsDERr0nAxEFCW/kZsOt7oZNJRkMxkyk/S0cqoYOapRBRkJ4fdzkQM9S1UUAeK0ctHNpAGnQZyRJSNrDPfJStQoLz4Z9n1Hb/krixAvPWAl48KsYEUmMDkJ3lvekJcqb7BP13eDG/szkBfA9p+RbC8q+HbVfupPG/4HdiFN1y+dRfg+z9z8wx+u8contPJh9QSwMEFAAAAAgAAAAhAPKfSdrpAAAASwIAAAsAAABfcmVscy8ucmVsc62SwU7DMAxA73xF5PuabkgIoaW7IKTdJjQ+wCRuG7WNo8SD7u+JkEAMjWkHjnHs52fL6808jeqNUvYcDCyrGhQFy86HzsDL/mlxDyoLBocjBzJwpAyb5mb9TCNKqcm9j1kVSMgGepH4oHW2PU2YK44Uyk/LaUIpz9TpiHbAjvSqru90+smA5oSpts5A2rolqP0x0jVsbltv6ZHtYaIgZ1r8yihkTB2JgXnU75yGV+ahKlDQ511W17v8PaeeSNChoLacaBFTqU7iy1q/dRzbXQnnz4xLQrf/uRyahYIjd1kJY/wy0ic30HwAUEsDBBQAAAAIAAAAIQCZcCPF3wAAALkCAAAaAAAAeGwvX3JlbHMvd29ya2Jvb2sueG1sLnJlbHO9ks9qwzAMh+97CqP7oiQbZZQ6vZRBr6N7AOMof2hiG0vtlrefaVm7Qik9lJ3MT0afPoQWy+9xUHuK3HunochyUOSsr3vXavjcvD+/gWIxrjaDd6RhIoZl9bT4oMFI6uGuD6wSxLGGTiTMEdl2NBrOfCCXfhofRyMpxhaDsVvTEpZ5PsP4lwHVBVOtaw1xXRegNlOge9i+aXpLK293Izm5MgK/fNxyRyQJamJLouFUYjw8RZaogNdlyn+WKW/JvDxSRlIvnUUO8Vi8uZDXRzqwTEO6rpPEMf+Ox4uLq34AUEsDBBQAAAAIAAAAIQBoRV4e8gcAAPMnAAAYAAAAeGwvd29ya3NoZWV0cy9zaGVldDEueG1sjZpbb9tIEoXf51cIel+K3c1rYDtILOtCYYHFZmbfGZm2hZFEL8nEM/9+KbOKNE8Vlc1DYJ1i9eVzd/VhWzef/zodZz+Lqj6U59u58fz5rDjvy8fD+fl2/sfvq38k81nd5OfH/Fiei9v530U9/3z3281bWf1ZvxRFM2sbONe385emef20WNT7l+KU1175WpzbyFNZnfKm/Vg9L+rXqsgf35NOx4X1/Whxyg/nedfCp+r/aaN8ejrsi2W5/3Eqzk3XSFUc86Ydfv1yeK3ndzePhzZ2mc+sKp5u51/Mp51z88XdzXvX/zkUb/WHn2dN/v1bcSz2TfHYApjPLjP7XpZ/XoLbVvIvqQuRu3of1b+q2WPxlP84Nv8u3zbF4fmlaRsJ+96WeZPf3VTl26x6b7x+/3//o27KU9fCRbi72V/iX/oHWvVwPh7OxbemaqOHttPm7vT6fLNo2rFcPi72lPT1etL+76OSdH896fFQvypZy+tZL1rOwy96qvJGyVpdz3rTctbXc/5bF3sla3M962et5Gyv5+QnJSe7nvNc5JWStfvFLzevvo+zFu1S69eb7RfW5aefd9bcLH5+XDqdHI3V+041kT/Wl6Qb0B863XnpWF9Rp15kx4E1d+AF0PWmi0AHW3p+rGadGozVHaojIK4H4nQgTgXiJoC4CSBuAgh16iVxCESopdjzgdXGqUScSsSpRFAdEQl6IkE3uHZ0wCTQWr3vVOPD08tOTx0gCQhJAjNfUbeew0VCHSReBNPcBNrkt6qaqYPf4bMjJGGPJKRF4kH+11BdJvR4iEjCiWUSEhNkuOKANbhOqKnUC2BIm1CFEmqrJ6P2AQq2MIIS9VAi/r3EAKULwFzuO9WJvUPN4E54oOc9nPqKAzjzNbck906k7h1VzSKVSadanUncM4mZCdaTWF0oMW01mOKSmvGRCT3vxdDQKmYmEFhThu9ZZBKr6yRWmcQqE2xhxCTpmSTdY4EH+V8TdZ0kE+uk022ATJJ+jwATDoSwPtc0oNBLcO8k6jpR1SxRmSTXamzaM0lpMqKgpGqNTQlhhHtt2UXwYH1Iae4RnjwcMBBY04hglptUXSWpSkQd+y69tnOMP1hef+rcoQhCIdkEPmYsKZTiWiG9PZAB2KqPiEo7jAuQbXgAAIdkpKPPYsfNT/D58EpguORb5GPU4sIJUexFyIdCFk9nCmiAOCLKLjeW4AbfcAQBGR2Q0QGhPAY0WFtjuf6LBaTbW07QANkpQHYSkJ0EZBmQWEGqod0a1f1mRje6Qh4DGqyuYfsqqg5FsBSTbONQ7jBqK0EvY9jz+jECIv8pImvTu14syEZ3vbqccd8ICOUxoMH5moDHgUcVRQSgYBpQMAUomATEkdghoH5g+J5EEQFIlTPuAQGhPAY0+GAT8tEpalCoAwqnAYVTgMJJQByJEwTEjSEe1fdudTkzuh8W8hjP4IgNOVBfbjDdE5McxBbhkN8UDpADHr5YrSjS/mrECdYb41Tw0Z2xLmdG98Y8iwk+gzs28SSfWOfTyYEwgxQQ70UPRrWqK5JDL7CiOsf9sYEW2ahueKvLmd71jucwQWfwyaY3yvhCZXSnTHIQCDrklZ3YWr1ZFoASAuTQYK95XG11FoB0w6zLmdEts7nqmc1gmg2ZVyuXj26bSY7FK+qSIlGEgFI+o/BVnCIWK9+aApdX8Rjx6O5ZlzN9Ejvx9PhmbzDQliys3F1WN9Akt7UZ8VAkxNcKbkhWH4oYLxIGmkOJh+1trO6gdTnTp7GzVx20HRy0JQsr7uO+WtV63pMcG89PP/5DWuS/sRZxq+Lde0UR4yVO0OptPppFq7tpXc70Ke3E02NYHy6K+abYCxGWakPvLVtmHy8/lhRKYwTElhkPeu7cCxDqmkO+5+N1oNXdtNXdNHeOgLCRMaDBTVvHXkgA0t00yQ5NyrJvSdyn8z2xuO/htsSWWtve5CdYjqxupnU5s7qZ5jFN8BnMtA2mvKLVzTTJzg+QTzDFhw2zuCPsI4HcYFMmf2N1L63LmdW9NMlTfAYvbcmyOvGyYXUvTbJDDEsKyKsx2/tlUa45gldga9tb/EDsL91M63JmdTNN8sRpbwczbSMuhGL96Gaa5MAXfKaumC1fJYvTvo8kwg7Z4Zo5REC6m9blzOpu2l69araDm7bkgeXbqkXLSYA6ORbnV6ybIW5GwUOeufUBAk9/BS7OL/3GWZczfQ478fSYzuCmLXngCNmg2yQ2SX96OcTThVKDeKiltsYgHurbM2J3JVydRXHWrTQnIB2y60gnubp2Bitt00mvqFtpktPwF1aIzKrBezPu0IvFWU8pXohJaw5Fci3pzlqXM5IFratX025w1s7nmog3ixTBUkSyC9EJUcBGwPfBsbMWXtH1N9N4bcbDCtCBbJx6Bb11uq8mGfE4tNtjPIOvdr1lxRdXp99McwLW1iUHRKV2fPuMf+lYUcSKhbXmxoRB2zj1AnrrdCdNssCDcxvjGZy0I0eJNpF0sXbIl6K9XXIAS+4DB7wwQDh9RMCxvHbwRtHpX71wqrnOSBZwcG5jOB++fsHfvxCFyKlfYrgn2eCf+5Yc8FPEw/fOBv9C6Hp/jfeJ3Fgi8ajfuNjqcqZPYud0E7348E2x1/y5+GdePR/O9exYPDW3c9+L57Oq+2LZ+89N+fr+UziffS+bpjzxp5cifyyqyyc3nz2VZcMfLt9f67+vd/c/UEsDBBQAAAAIAAAAIQB5Ur3rzw4AAPtnAAAYAAAAeGwvd29ya3NoZWV0cy9zaGVldDIueG1spZ3bbiO5EYbv8xSG72M3z+TA48WOuxcJkACLbA7XWlseC2tbhqTZSd4+ksVqcqr4zza6bwyryy39qi7yY/FQvvnhvy/PF7+vd/vN9vXjpbrqLi/Wr/fbh83r54+X//rnT3+Olxf7w+r1YfW8fV1/vPzfen/5w+2fbr5ud7/tn9brw8XxDV73Hy+fDoe3D9fX+/un9ctqf7V9W78eLY/b3cvqcHy5+3y9f9utVw/vN708X+uu89cvq83r5fkdPuymvMf28XFzv+63919e1q+H85vs1s+rw1H+/mnztr+8vXnYHG2n73OxWz9+vPxRfRiUU5fXtzfvn/3vzfrrvvr94vRVft1ufzu9+OvDx8vu9KfX4m9/epfx8+7iYf24+vJ8+Mf261/Wm89Ph6Pb3Pju/eqwur3Zbb9eHL+POvru/ef9l/1h+3J+h9OF25v7k/3H8Q+OVzevz5vX9S+H3dG6OX7o4faX9dvq+epv69fPh6eb68NR1On69X2++9OUu/+zeWjefPf9m39eH7730f2Uu9FHD3+g+219v1nvv73x+ujR0a169N/pt99v3ZW6uf699sz5urly316/O19XV/bb6/35enelv70+aKBxvz5s96vvSDSjRPP+1vYqMYnn64YJNECgAQLNXIF2FGizwMAE2uxD9pF3Nktk0nsLJNq5Et0o0WWJnkl0WSJ7/HcuS2SPv3dAopsr0Y8S/TkSmUCfBTLhdx48Zg8E+rkCwygw5KbCPvJTyBJZfN6FLJGFRR+yRPY+Q5grMY4SI3jMMUtkH3kXgRdjlsgidIhzJaZRYmo+5gQEJhCHCTzmNFeg6gpSuuxF/qCzQYsnnQ3Sj9kgdR4Nc4VW7FOgZ8wG2ayzQfozGzp+x6AQZP5YaIGM0qDpKMJM4EI1EopAo2aTRhXUKGJN5EINCFBFvPFcKAKOmk0cVZCjLBJqm1RUFkWoRQ9+NnVUwY4i7hgu07VlEncUl+mQzNnkUQU9KrNHevNsEA+d6KO5TIQfNZs/qgBIEYH4QCMbRI91pwJqRghCajaFVMGQiqi9E4hED0ok4mMiFZHQ2SxSBUYqgeGvIiDx8a9KqCERkjgz1Wwm6cIk3YFHnw2Gh+6dJibx4YfugFA9m0m6MEkrlFAoJBQxSSskdDaTdJX4ICZpYhJvTFojj8LkZzaTdGGSNsijxCQOT22QRw1oTHo2k3RhkiYm8dGnpkSID+KzgXf2GiVCejaTdGGSdsiflAxxKmmHHjxRyXGhs6mkC5W0B4zXHkUocSlxoYhLejaXdOGSDs0xfb4svElM4iMmHZDI2UzShUk6tkWi1EhHJBMRSc8mki5E0kQkzYUiImmUI2mUJOnZRDKFSKYDQk0HPGpQlmRQlmRmE8kUIhnKkjg6DRGJz88YBR69UUjobCKZQiSjQXs3RCSezhmNhCIimfnzcdWEHBGJo9OgLMkgIhlEJDObSKYQyVgUo4Qq4VGLhKI8ycxmkilMMsQk3j0ZSqBEjDrUmNAUnZnNJFOYZIhJYjLWoxj1yKOISWY2k0xhkmkzydBknfAnUYmndAZRycymkilUMhE9eOIS7+4NypRMREJnc8kULpmEHjxxiQ/uDMqUTEJNaTaXbOGSRbN32cCHI7YD/rSISnY2lWyhkkV5kiUq8T7UojzJIirZ2VSyhUpWNxuSRQtEViN/EpN4OmdnM8kWJlmauRNCzwYtRvXWIKEGCZ2/VlQtFlkUn3C5CK4XwQWj2UyyhUnWtR+8Qw+eiMQHI5aI5LnM2USyhUjWo4ZEROITDhZlSZaIxAcjdjaRbCGSpUk6Pryz7TzJEpF4/2kDis/ZRLKFSDYif0bkT5QpWUQkO5tIthDJEpH4dINNqCEhIlmUKdnZRHKFSI4yJT4J7ihT4hM4jpgkll8Rk9xsJrnCJKeaLd4RkXiEOgX86RCR3GwiuUIkdwYJT+ecBo/dUV7Fp28c2MwwOESk920+99vn7e57UguV3BklXnT2ziCxgGO9AynU4BCVpoktZHI2i+VDPEdkErsEiGW8J3UgjRocItM0sdWGBpQxZYPkfTaIWCVs8d7UITpNk1oI5Xz2q5Dqs1TeozrKsniP6kAyNThEqGliC6UcWmPKhobYgCI2IM8iTk0TW1jlYvas6FmJVSIIaIeE6AsAxAaHWDVNbOGVQxlUNmgx4nep/S36bOD5k0O0miTVF2L5LvuVo9WjHRC+AxHrQXo1eESsaWILtTxlUnySJxu0wKsnoPFuywOgDR5xa5rYwi7fzqbyZQ4DD5KsPht4APhF3PKFW55m+His+vamO0/U4rMnHlHLL6KWL9TymVpcqM0PX/jUNjHgm0tRg1/EK1945V1uUnxY7R1qUjT1x7sqD6b+Br+IWL7aikc5lfCqR2LBPr3eE7FE+19ELF+I5UP2LCeWp7k+PmzxYLtE70HKNfhFxPKFWD4iz8Z2wyJeiR4gooa1iFe+8MrTShTPVz3xSvSrRDi+oOvbvPKLeBUKrwLxiiMgEK94LxA64NcAUq8hLOJVKLwKCoRAIF7xmZWAeBWIV9yzYRGvQuFVoF0SHAQBZVsBrGL1gbKtyMUuolYo1AqUbfE+NtAcIB+7ZgPvBwKYGhzCImqFQq1AuRYfuQaLggDlWgHlWmERu0JhV0DsCg75FbErELv4zEBYxK5Q2BUo2xLbtBG7AmVbfJgdELvCInaFaks5sUv0Be05wYDIFRC5wiJyhUKuQLkWx0E2NIKA2CV6ArDtfAiL2BUKu0ICY4JsEH5t7krvA00YBi50EbliIVfsmqPXiPKsiLgVEbfiIm7Fwq1I3OI+jcQtvgIYiVtCrGqOB+IiasVCrUjU4nlWNshUOxLOeKxGopYQu4hasVArGiTWILFg41+fDULqImrFQq1I+yl4HxCJWnxIGGnqkFMrErU4COIiasVCrejaTcshqWBfYB/BwtYQFzErFmZFyrc4s7KBd1eRiCUaFpofjIuIFQuxYmh7lXItEatodjCCbYBDXESsWB2GImKJDotmB3liGNHsYETZVlxErFiIFYlYfEAYKdsSQUDZlji8RdQSp7cWUSsVaiVa0eJhkNp7LBJRizethGYH0yJqpUKtRNQSJ85QtpUN3KtJIamLqJUKtRKiViJqccAmDUIgEbXE+bhF1EqFWgnlWqk9Q5hohpBPECVwYHdIi6iVCrUSolayyK/tOcKEmJUWMSsVZqX2nouE1rQS2LLeZwMfCaRFxEqFWAnNECbKsjheEzGLTxEnlGWlRcxKhVkJrWmldpaVApIKNgkOaRGzUmFWikgqZVk8I0gRiQUbBYe0iFmpOspLzOLTbikhscQsEa+IWWkRs1RXH+slaonDXh2lW+K0V9cEWk/XZVpwsiwSXB3v7VDGRRaZc5NFopYs0sUnyyLJupKswTCGLHJJdrxHCNbtpzKcLEDwZvd587q5//5J7+q8b0fJlzj52VH2JY4mj/dIFxPLklAMYTZJcXXwtztDKDSiGBz97Yh/PAcji25EMSTaJL3VCeDOwYhAK19kEXTpySInkU+WJYqrw8Ad2q9BFulhcH64J4tYgBhOliV6q1PBXcgRwVlMFqkXTDv2ZGlFBETcJL3V4eCOiCXKFnQEOdkPo3Uwssg5upNlieLqlHCXsodlDCPSkUXEfU+WVgxD2E1RXJewULQiJtChIOtUh6J4rG8hFMM6FtMU17UsVPYxH0yosZoFH6aTRSQhPVka7IAFLaYprliniHWinxjLWvDFMTVWwpCKm1tABgXrWkzTW7FOoc2IZGmwThHrRBQryDpY4GKa4op1ilbJBJ1RmQtFrJMFTiDrYKWLaXor1inilmx1Dra6ZsrXj3cItYs4Vxe9wFUvFNqYWO6R8Uv3iEINsPbFNMUV6VSAEYzOdamxaIaMYGKdGBDDIhjTFFesUzQHKfuI9oYPNdbOkBFBpJO98CLS1fUwFJFORjBtRpQxARbberI0Rj+wKMYkxVVhDJXLWTQUayKd4IYmOooR8XiPiAlYHWOa4op0ua6FGK1p1fbWHVlkLSRwJGxQsETGNLUV5bRu++qT0pByY20NERFaQ/8u4lxVL0ONBTOkh9E+ELLI3RXjPULvIspVZTOUtjCCLdRrUZuje4TeRZSr6mcoDTM6jdbWyNLyL8zoYBGNaYor0mnf9tYnNZbSEP2w9jCGiXRiJAGraUxTXJFOU04nRsMakk43l+Z6NRbckB5exLmquIbSxDmpF20PIYvcH6I0JB2ssTFNcUU6TdQS8yi6vUeErrdiGGZ0sNTGJL1VuQ1liFli7GM65GED1un6co+IYVhzY5riinMGZnRUeUPopQlPERGoJMegYOmNaXor0uWiGaHhYQ09rNtR1KuxNofIj2ANjmmK62qFJisWbB4rcUjFIAvsySLrFS4iXVWOQxkLYxiSbqzhIWMYbPEfFKzJMU1xxToDWWfQ3kdl0EYSNRbtEOM1WJxjmuKKdca3I/ITWRrjYQNW83o1Vu8QWR2s0jFNccU6Q6wTdDZg/tIE2OpgTgfLdUzTW7HOoFNnaizaIauEgn3/PVkaeTOs2zFNccU6k2C7o6xOVIo1tP1fZB0G0g4W8JikuCrioWx7X6QCRTzUWPVDkMPCuUtYx2Oa2op0luYhRT88VvMQ/h0LgMiKrJQHypKsi1hXFfVQlrgl2txY2EMq1iiGLdhWOShY22Oa4op11kAfG6gYrtShqiCDgkU+pimuS/SifSdkaeRJ4z1SMW0+EeMJWO1jmuKKdpZoJxXTRhORdVCZEDkqtmDXyqBg4Y9piivaWZjZWZjZWZjZWSKk4DOsADJNcUU7i46rKVAHhK6LA649WVoeXkS7qhyIspB2Fq7WUQ0RoTfCGF7EuqoqiLLEOjHjakFeZ0HJxZ4sYnwJy4JMUluVBlEOnVxTY3EQMZYY64kIcjg4ewnrg0xTXLGOioTIftiBrG68Qxb1pqxOsBkWCvmu3uvqH8i8rT6v/746/fH+4nn9eHyz7urYKHbn/zfz/vth+/b+27E3+nV7OGxf6NXTevWw3p1eHWHwuN0e6MXp39qM/7fn9v9QSwMEFAAAAAgAAAAhALn+GMxUAQAAVAIAAA8AAAB4bC93b3JrYm9vay54bWyNUU1PwzAMvfMrotxZ26ir2LR2El9iEgIkxnYOjbtGS5MqSen273E6dQxx4WQ/5/nZz1ksD40iX2CdNDqnySSmBHRphNS7nH6sH69vKHGea8GV0ZDTIzi6LK4WvbH7T2P2BPu1y2ntfTuPIlfW0HA3MS1ofKmMbbhHaHeRay1w4WoA36iIxXEWNVxqelKY2/9omKqSJdybsmtA+5OIBcU9bu9q2TpaLCqpYHMyRHjbvvAG1z4oShR3/kFIDyKnKULTw6+C7drbTqoApvGURsXZ5JslAireKb/G1UZ1PBdLGcsCM7A2Enr30xQgOWylFqbPKUvxsscRJVNK+iHfSuFrLGTx7Fx7ArmrfU5nWRYH8ehCfbjfGIkezL2HPMF/CnGF+2Nu5xITuxJJUPjDZhdsdsFmw7xxSMlVid5DGGRZOk1mlFSdUndYe9XPhg/jQtPou/gGUEsDBBQAAAAIAAAAIQA/2O8hsQUAAFMbAAATAAAAeGwvdGhlbWUvdGhlbWUxLnhtbO1ZTY/TRhi+8ytGvoPjxA7ZFVm0ySbQwsJqN1BxnNgTe8jYY81MdsmtgmOlSlVp1Uul3nqo2iKB1Av9NdtStVTiL/T1R5LxZrJkYasWQQ6JZ/y83x9+x7ly9UHM0CERkvKkbTmXahYiic8DmoRt686gf7FlIalwEmDGE9K2pkRaV7cuXMGbKiIxQUCeyE3ctiKl0k3blj5sY3mJpySBeyMuYqxgKUI7EPgI2MbMrtdqTTvGNLFQgmPgens0oj5Bg4yltTVj3mPwlSiZbfhMHPi5RJ0ixwZjJ/uRU9llAh1i1rZATsCPBuSBshDDUsGNtlXLP5a9dcWeEzG1glaj6+efkq4kCMb1nE6Ewzmh03c3Lu/M+dcL/su4Xq/X7TlzfjkA+z5Y6ixh3X7L6cx4aqDicpl3t+bV3Cpe499Ywm90Oh1vo4JvLPDuEr5Va7rb9QreXeC9Zf07291us4L3FvjmEr5/eaPpVvE5KGI0GS+hs3jOIzOHjDi7boS3AN6aJcACZWvZVdAnalWuxfg+F30A5MHFiiZITVMywj7gujgeCoozAXiTYO1OseXLpa1MFpK+oKlqWx+nGCpiAXn1/MdXz5+iV8+fHD98dvzwl+NHj44f/mwgvI6TUCd8+f0Xf3/7Kfrr6XcvH39lxksd//tPn/3265dmoNKBL75+8sezJy+++fzPHx4b4NsCD3X4gMZEolvkCO3zGGwzCCBDcTaKQYRphQJHgDQAeyqqAG9NMTPhOqTqvLsCGoAJeG1yv6LrQSQmihqAN6K4AtzlnHW4MJpzI5OlmzNJQrNwMdFx+xgfmmR3T4S2N0khk6mJZTciFTX3GEQbhyQhCmX3+JgQA9k9Sit+3aW+4JKPFLpHUQdTo0sGdKjMRNdpDHGZmhSEUFd8s3sXdTgzsd8hh1UkFARmJpaEVdx4DU8Ujo0a45jpyJtYRSYlD6bCrzhcKoh0SBhHvYBIaaK5LaYVdW9g6ETGsO+yaVxFCkXHJuRNzLmO3OHjboTj1KgzTSId+5EcQ4pitMeVUQlerZBsDXHAycpw36VEna2s79AwMidIdmciyq5d6b8xTU5rxoxCN/7QjGfwbXg0mUriZAtehXsHG+8OniR7BHL9Q9/90Hffx767qpbX7baLBmvrc3HOL145JI8oYwdqyshNmbdmCUoHfdjMFznRfCZPI7gsxVVwocD5NRJcfUJVdBDhFMQ4uYRQlqxDiVIu4SRgreSdHycpGJ/vebMzIKCx2uVBsd3Qz4ZzNvkqlLqgRsZgXWGNy28nzCmAa0pzPLM071RptuZNqAaEs4O/06wXoiFjMCNB5veCwSws5x4iGeGAlDFyjIY4jTXd1nq91zRpG423k7ZOkHRx7gpx3jlEqbYUJXu5HFlSXaEj0Mqrexbycdq2RjBJwWWcAj+ZNSDMwqRt+ao05bXFfNJgc1o6tZUGV0SkQqodLKOCKr81e3WSLPSve27mh/MxwNCN1tOi0XL+Qy3sk6EloxHx1YqdxbK8xyeKiIMoOEJDNhH7GPR2i+wKqIRnRn22EFChbpl41covq+DkK5qyOjBLI1z2pJYW+wKeX891yFeaevYK3d/QlMY5muK9v6ZkmQtjayPID1QwBgiMshxtW1yoiEMXSiPq9wUMDrks0AtBWWQqIZa9b850JYeLvlXwKJpcGKl9GiJBodOpSBCyp0o7X8PMqevP1xmjss/M1ZVp8Tskh4QNsuptZvZbKJp1k9IROe5k0GxTdQ3D/v948nFXTD6njwcLQe5ZZhFXa/rao2Dj7VQ446O2bra47q39qE3h8IGyL2jcVPhsMd8O+D5EH80nSgSJeLFVlt98cwg6tzTjMlb/7hi1CEFrRbzPc/jUnN1Y4ezTxb25sz2Dr73TXW0vl6itHWTy1dIfT3x4H2TvwEFpwpQs3iY9gKNmd/aXAfCxF6RbF/4BUEsDBBQAAAAIAAAAIQCvzhJdwwEAAFIEAAANAAAAeGwvc3R5bGVzLnhtbL1U34ucMBB+718R8t5zFXrcFfUohYVCWw5uC32NGjUwSSQZF72//iZG3V1oOehDX8zMNzPf/HCS/GnSwM7SeWVNwdO7A2fS1LZRpiv4r9Px4wNnHoVpBFgjCz5Lz5/KD7nHGeRLLyUyYjC+4D3i8DlJfN1LLfydHaQhS2udFkiq6xI/OCkaH4I0JNnhcJ9ooQwv89Ya9Ky2o8GCZytQ5v6VnQVQWSlPyry2YB1Doqc6FsQILaPHVwGqciqArdAK5ghnAVgqWv20MtYFMIkZ4rdK/keu5fAUpABumyWgzAeBKJ05ksJW+TQPlN7Q4CPN4veOd+fEnGafrgKWg/JW1jX0o7fMKd+gMgfZIgU41fXhRDskwYhoNQmNEp01AgLlFrEKRFtLgJewDb/bG+6pZWbUR43fmoLTWoXuN5EKWsVIE5XAf80Wua9os3+iZVO78/8tOn0/molhgPloY39R+wKqM1puLYtNZb116pVcwx7UBMi4CFO7trh3t/R6M7cdZWHlCv4z3CC4qqMaFaAyf5gZcTbTZVyLFUVFF/UmC3E0shUj4Gk3Fvwi/5CNGvXj7vWszhZXr4v8PSxLer9UcHkNyjdQSwMEFAAAAAgAAAAhAG5Nq10lAQAAUAIAABEAAABkb2NQcm9wcy9jb3JlLnhtbJ2Sy2rDMBBF9/0Ko70t2e4jGNuBtmTVQKEpLd0JaeKIWg8ktY7/vrKdOAlkVdBGunfO3BlULveyjX7BOqFVhdKEoAgU01yopkLvm1W8QJHzVHHaagUV6sGhZX1TMlMwbeHVagPWC3BRAClXMFOhnfemwNixHUjqkuBQQdxqK6kPV9tgQ9k3bQBnhNxjCZ5y6ikegLGZieiA5GxGmh/bjgDOMLQgQXmH0yTFJ68HK93VglE5c0rhewNXrUdxdu+dmI1d1yVdPlpD/hR/rl/exlFjoYZVMUB1yVnBLFCvbV3i80tYXEudX4cVbwXwxz7oV94Og0x1wKMQoJjiHpWP/Ol5s0J1RjISk9s4vduQhyJfhPM1tLyoPwHlocm/iUfAlPvyE9R/UEsDBBQAAAAIAAAAIQAr2hwKewEAACsDAAAQAAAAZG9jUHJvcHMvYXBwLnhtbJ2SQWvjMBCF7/srjO6NHLOUEmSVknbJoWUDSduzVh7HIrIkNFOT7K9f2SGu0xwK69PTzOPp82jE/aG1WQcRjXclm89yloHTvjJuV7LX7a+bO5YhKVcp6x2U7AjI7uUPsY4+QCQDmKUEhyVriMKCc9QNtApnqe1Sp/axVZSOccd9XRsNj15/tOCIF3l+y+FA4CqobsIYyE6Ji47+N7TyuufDt+0xpDwpHkKwRitKPylfjI4efU3Z00GDFXzaFCloA/ojGjrKXPDpUWy0srBMwbJWFkHwz4JYgepntlYmohQdLTrQ5GOG5m+aWsGyPwqhxylZp6JRjtjJdjoM2gakKN993GMDQCj4WBzk1DvV5qcsBkMSl0Y+giR9ibg1ZAF/12sV6TvigYFNGDc933zKd9kqrtDPEF+uXfo2KJdmy0f1bNweX8PWPyqC86Qvi2LTqAhVepzxJcaCWCXkaHv/slFuB9XZc93o9+LttPtyXszy9A3rcK4J/rnm8h9QSwECAAAUAAAACAAAACEAng4nlkIBAACPBAAAEwAAAAAAAAABAAAAAAAAAAAAW0NvbnRlbnRfVHlwZXNdLnhtbFBLAQIAABQAAAAIAAAAIQDyn0na6QAAAEsCAAALAAAAAAAAAAEAAAAAAHMBAABfcmVscy8ucmVsc1BLAQIAABQAAAAIAAAAIQCZcCPF3wAAALkCAAAaAAAAAAAAAAEAAAAAAIUCAAB4bC9fcmVscy93b3JrYm9vay54bWwucmVsc1BLAQIAABQAAAAIAAAAIQBoRV4e8gcAAPMnAAAYAAAAAAAAAAEAAAAAAJwDAAB4bC93b3Jrc2hlZXRzL3NoZWV0MS54bWxQSwECAAAUAAAACAAAACEAeVK9688OAAD7ZwAAGAAAAAAAAAABAAAAAADECwAAeGwvd29ya3NoZWV0cy9zaGVldDIueG1sUEsBAgAAFAAAAAgAAAAhALn+GMxUAQAAVAIAAA8AAAAAAAAAAQAAAAAAyRoAAHhsL3dvcmtib29rLnhtbFBLAQIAABQAAAAIAAAAIQA/2O8hsQUAAFMbAAATAAAAAAAAAAEAAAAAAEocAAB4bC90aGVtZS90aGVtZTEueG1sUEsBAgAAFAAAAAgAAAAhAK/OEl3DAQAAUgQAAA0AAAAAAAAAAQAAAAAALCIAAHhsL3N0eWxlcy54bWxQSwECAAAUAAAACAAAACEAbk2rXSUBAABQAgAAEQAAAAAAAAABAAAAAAAaJAAAZG9jUHJvcHMvY29yZS54bWxQSwECAAAUAAAACAAAACEAK9ocCnsBAAArAwAAEAAAAAAAAAABAAAAAABuJQAAZG9jUHJvcHMvYXBwLnhtbFBLBQYAAAAACgAKAIQCAAAXJwAAAAA=" download="mtcars and iris datasets.xlsx">
#>   <button class="btn btn-warning"><i class="fa fa-save"></i> Download data</button>
#> </a>

# Passing a named list with data frames to the function
list('mtcars dataset' = mtcars, 'iris dataset' = iris) %>%
  download_this(
    output_name = "mtcars and iris datasets",
    output_extension = ".csv",
    button_label = "Download data",
    button_type = "warning",
    has_icon = TRUE,
    icon = "fa fa-save"
  )

#> <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css" integrity="sha256-h20CPZ0QyXlBuAw7A+KluUYx/3pK+c7lYEpqLTlxjYQ=" crossorigin="anonymous" />
#> <a href="data:application/vnd.openxmlformats-officedocument.spreadsheetml.sheet;base64,UEsDBBQAAAAIAAAAIQCeDieWQgEAAI8EAAATAAAAW0NvbnRlbnRfVHlwZXNdLnhtbLWUy27CMBBF9/2KyNsqNrCoqorAoo9ly4J+gGtPiIVf8hgKf99JQpFaUSii3cSK5849dyzL4+nG2WINCU3wFRvyASvAq6CNX1Tsdf5U3rICs/Ra2uChYltANp1cjefbCFhQs8eKNTnHOyFQNeAk8hDBU6UOyclMv2kholRLuQAxGgxuhAo+g89lbj3YZPwAtVzZXDxuaLsPksAiK+57YcuqmIzRGiUz1cXa62+Uckfg1NlpsDERr0nAxEFCW/kZsOt7oZNJRkMxkyk/S0cqoYOapRBRkJ4fdzkQM9S1UUAeK0ctHNpAGnQZyRJSNrDPfJStQoLz4Z9n1Hb/krixAvPWAl48KsYEUmMDkJ3lvekJcqb7BP13eDG/szkBfA9p+RbC8q+HbVfupPG/4HdiFN1y+dRfg+z9z8wx+u8contPJh9QSwMEFAAAAAgAAAAhAPKfSdrpAAAASwIAAAsAAABfcmVscy8ucmVsc62SwU7DMAxA73xF5PuabkgIoaW7IKTdJjQ+wCRuG7WNo8SD7u+JkEAMjWkHjnHs52fL6808jeqNUvYcDCyrGhQFy86HzsDL/mlxDyoLBocjBzJwpAyb5mb9TCNKqcm9j1kVSMgGepH4oHW2PU2YK44Uyk/LaUIpz9TpiHbAjvSqru90+smA5oSpts5A2rolqP0x0jVsbltv6ZHtYaIgZ1r8yihkTB2JgXnU75yGV+ahKlDQ511W17v8PaeeSNChoLacaBFTqU7iy1q/dRzbXQnnz4xLQrf/uRyahYIjd1kJY/wy0ic30HwAUEsDBBQAAAAIAAAAIQCZcCPF3wAAALkCAAAaAAAAeGwvX3JlbHMvd29ya2Jvb2sueG1sLnJlbHO9ks9qwzAMh+97CqP7oiQbZZQ6vZRBr6N7AOMof2hiG0vtlrefaVm7Qik9lJ3MT0afPoQWy+9xUHuK3HunochyUOSsr3vXavjcvD+/gWIxrjaDd6RhIoZl9bT4oMFI6uGuD6wSxLGGTiTMEdl2NBrOfCCXfhofRyMpxhaDsVvTEpZ5PsP4lwHVBVOtaw1xXRegNlOge9i+aXpLK293Izm5MgK/fNxyRyQJamJLouFUYjw8RZaogNdlyn+WKW/JvDxSRlIvnUUO8Vi8uZDXRzqwTEO6rpPEMf+Ox4uLq34AUEsDBBQAAAAIAAAAIQBoRV4e8gcAAPMnAAAYAAAAeGwvd29ya3NoZWV0cy9zaGVldDEueG1sjZpbb9tIEoXf51cIel+K3c1rYDtILOtCYYHFZmbfGZm2hZFEL8nEM/9+KbOKNE8Vlc1DYJ1i9eVzd/VhWzef/zodZz+Lqj6U59u58fz5rDjvy8fD+fl2/sfvq38k81nd5OfH/Fiei9v530U9/3z3281bWf1ZvxRFM2sbONe385emef20WNT7l+KU1175WpzbyFNZnfKm/Vg9L+rXqsgf35NOx4X1/Whxyg/nedfCp+r/aaN8ejrsi2W5/3Eqzk3XSFUc86Ydfv1yeK3ndzePhzZ2mc+sKp5u51/Mp51z88XdzXvX/zkUb/WHn2dN/v1bcSz2TfHYApjPLjP7XpZ/XoLbVvIvqQuRu3of1b+q2WPxlP84Nv8u3zbF4fmlaRsJ+96WeZPf3VTl26x6b7x+/3//o27KU9fCRbi72V/iX/oHWvVwPh7OxbemaqOHttPm7vT6fLNo2rFcPi72lPT1etL+76OSdH896fFQvypZy+tZL1rOwy96qvJGyVpdz3rTctbXc/5bF3sla3M962et5Gyv5+QnJSe7nvNc5JWStfvFLzevvo+zFu1S69eb7RfW5aefd9bcLH5+XDqdHI3V+041kT/Wl6Qb0B863XnpWF9Rp15kx4E1d+AF0PWmi0AHW3p+rGadGozVHaojIK4H4nQgTgXiJoC4CSBuAgh16iVxCESopdjzgdXGqUScSsSpRFAdEQl6IkE3uHZ0wCTQWr3vVOPD08tOTx0gCQhJAjNfUbeew0VCHSReBNPcBNrkt6qaqYPf4bMjJGGPJKRF4kH+11BdJvR4iEjCiWUSEhNkuOKANbhOqKnUC2BIm1CFEmqrJ6P2AQq2MIIS9VAi/r3EAKULwFzuO9WJvUPN4E54oOc9nPqKAzjzNbck906k7h1VzSKVSadanUncM4mZCdaTWF0oMW01mOKSmvGRCT3vxdDQKmYmEFhThu9ZZBKr6yRWmcQqE2xhxCTpmSTdY4EH+V8TdZ0kE+uk022ATJJ+jwATDoSwPtc0oNBLcO8k6jpR1SxRmSTXamzaM0lpMqKgpGqNTQlhhHtt2UXwYH1Iae4RnjwcMBBY04hglptUXSWpSkQd+y69tnOMP1hef+rcoQhCIdkEPmYsKZTiWiG9PZAB2KqPiEo7jAuQbXgAAIdkpKPPYsfNT/D58EpguORb5GPU4sIJUexFyIdCFk9nCmiAOCLKLjeW4AbfcAQBGR2Q0QGhPAY0WFtjuf6LBaTbW07QANkpQHYSkJ0EZBmQWEGqod0a1f1mRje6Qh4DGqyuYfsqqg5FsBSTbONQ7jBqK0EvY9jz+jECIv8pImvTu14syEZ3vbqccd8ICOUxoMH5moDHgUcVRQSgYBpQMAUomATEkdghoH5g+J5EEQFIlTPuAQGhPAY0+GAT8tEpalCoAwqnAYVTgMJJQByJEwTEjSEe1fdudTkzuh8W8hjP4IgNOVBfbjDdE5McxBbhkN8UDpADHr5YrSjS/mrECdYb41Tw0Z2xLmdG98Y8iwk+gzs28SSfWOfTyYEwgxQQ70UPRrWqK5JDL7CiOsf9sYEW2ahueKvLmd71jucwQWfwyaY3yvhCZXSnTHIQCDrklZ3YWr1ZFoASAuTQYK95XG11FoB0w6zLmdEts7nqmc1gmg2ZVyuXj26bSY7FK+qSIlGEgFI+o/BVnCIWK9+aApdX8Rjx6O5ZlzN9Ejvx9PhmbzDQliys3F1WN9Akt7UZ8VAkxNcKbkhWH4oYLxIGmkOJh+1trO6gdTnTp7GzVx20HRy0JQsr7uO+WtV63pMcG89PP/5DWuS/sRZxq+Lde0UR4yVO0OptPppFq7tpXc70Ke3E02NYHy6K+abYCxGWakPvLVtmHy8/lhRKYwTElhkPeu7cCxDqmkO+5+N1oNXdtNXdNHeOgLCRMaDBTVvHXkgA0t00yQ5NyrJvSdyn8z2xuO/htsSWWtve5CdYjqxupnU5s7qZ5jFN8BnMtA2mvKLVzTTJzg+QTzDFhw2zuCPsI4HcYFMmf2N1L63LmdW9NMlTfAYvbcmyOvGyYXUvTbJDDEsKyKsx2/tlUa45gldga9tb/EDsL91M63JmdTNN8sRpbwczbSMuhGL96Gaa5MAXfKaumC1fJYvTvo8kwg7Z4Zo5REC6m9blzOpu2l69araDm7bkgeXbqkXLSYA6ORbnV6ybIW5GwUOeufUBAk9/BS7OL/3GWZczfQ478fSYzuCmLXngCNmg2yQ2SX96OcTThVKDeKiltsYgHurbM2J3JVydRXHWrTQnIB2y60gnubp2Bitt00mvqFtpktPwF1aIzKrBezPu0IvFWU8pXohJaw5Fci3pzlqXM5IFratX025w1s7nmog3ixTBUkSyC9EJUcBGwPfBsbMWXtH1N9N4bcbDCtCBbJx6Bb11uq8mGfE4tNtjPIOvdr1lxRdXp99McwLW1iUHRKV2fPuMf+lYUcSKhbXmxoRB2zj1AnrrdCdNssCDcxvjGZy0I0eJNpF0sXbIl6K9XXIAS+4DB7wwQDh9RMCxvHbwRtHpX71wqrnOSBZwcG5jOB++fsHfvxCFyKlfYrgn2eCf+5Yc8FPEw/fOBv9C6Hp/jfeJ3Fgi8ajfuNjqcqZPYud0E7348E2x1/y5+GdePR/O9exYPDW3c9+L57Oq+2LZ+89N+fr+UziffS+bpjzxp5cifyyqyyc3nz2VZcMfLt9f67+vd/c/UEsDBBQAAAAIAAAAIQB5Ur3rzw4AAPtnAAAYAAAAeGwvd29ya3NoZWV0cy9zaGVldDIueG1spZ3bbiO5EYbv8xSG72M3z+TA48WOuxcJkACLbA7XWlseC2tbhqTZSd4+ksVqcqr4zza6bwyryy39qi7yY/FQvvnhvy/PF7+vd/vN9vXjpbrqLi/Wr/fbh83r54+X//rnT3+Olxf7w+r1YfW8fV1/vPzfen/5w+2fbr5ud7/tn9brw8XxDV73Hy+fDoe3D9fX+/un9ctqf7V9W78eLY/b3cvqcHy5+3y9f9utVw/vN708X+uu89cvq83r5fkdPuymvMf28XFzv+63919e1q+H85vs1s+rw1H+/mnztr+8vXnYHG2n73OxWz9+vPxRfRiUU5fXtzfvn/3vzfrrvvr94vRVft1ufzu9+OvDx8vu9KfX4m9/epfx8+7iYf24+vJ8+Mf261/Wm89Ph6Pb3Pju/eqwur3Zbb9eHL+POvru/ef9l/1h+3J+h9OF25v7k/3H8Q+OVzevz5vX9S+H3dG6OX7o4faX9dvq+epv69fPh6eb68NR1On69X2++9OUu/+zeWjefPf9m39eH7730f2Uu9FHD3+g+219v1nvv73x+ujR0a169N/pt99v3ZW6uf699sz5urly316/O19XV/bb6/35enelv70+aKBxvz5s96vvSDSjRPP+1vYqMYnn64YJNECgAQLNXIF2FGizwMAE2uxD9pF3Nktk0nsLJNq5Et0o0WWJnkl0WSJ7/HcuS2SPv3dAopsr0Y8S/TkSmUCfBTLhdx48Zg8E+rkCwygw5KbCPvJTyBJZfN6FLJGFRR+yRPY+Q5grMY4SI3jMMUtkH3kXgRdjlsgidIhzJaZRYmo+5gQEJhCHCTzmNFeg6gpSuuxF/qCzQYsnnQ3Sj9kgdR4Nc4VW7FOgZ8wG2ayzQfozGzp+x6AQZP5YaIGM0qDpKMJM4EI1EopAo2aTRhXUKGJN5EINCFBFvPFcKAKOmk0cVZCjLBJqm1RUFkWoRQ9+NnVUwY4i7hgu07VlEncUl+mQzNnkUQU9KrNHevNsEA+d6KO5TIQfNZs/qgBIEYH4QCMbRI91pwJqRghCajaFVMGQiqi9E4hED0ok4mMiFZHQ2SxSBUYqgeGvIiDx8a9KqCERkjgz1Wwm6cIk3YFHnw2Gh+6dJibx4YfugFA9m0m6MEkrlFAoJBQxSSskdDaTdJX4ICZpYhJvTFojj8LkZzaTdGGSNsijxCQOT22QRw1oTHo2k3RhkiYm8dGnpkSID+KzgXf2GiVCejaTdGGSdsiflAxxKmmHHjxRyXGhs6mkC5W0B4zXHkUocSlxoYhLejaXdOGSDs0xfb4svElM4iMmHZDI2UzShUk6tkWi1EhHJBMRSc8mki5E0kQkzYUiImmUI2mUJOnZRDKFSKYDQk0HPGpQlmRQlmRmE8kUIhnKkjg6DRGJz88YBR69UUjobCKZQiSjQXs3RCSezhmNhCIimfnzcdWEHBGJo9OgLMkgIhlEJDObSKYQyVgUo4Qq4VGLhKI8ycxmkilMMsQk3j0ZSqBEjDrUmNAUnZnNJFOYZIhJYjLWoxj1yKOISWY2k0xhkmkzydBknfAnUYmndAZRycymkilUMhE9eOIS7+4NypRMREJnc8kULpmEHjxxiQ/uDMqUTEJNaTaXbOGSRbN32cCHI7YD/rSISnY2lWyhkkV5kiUq8T7UojzJIirZ2VSyhUpWNxuSRQtEViN/EpN4OmdnM8kWJlmauRNCzwYtRvXWIKEGCZ2/VlQtFlkUn3C5CK4XwQWj2UyyhUnWtR+8Qw+eiMQHI5aI5LnM2USyhUjWo4ZEROITDhZlSZaIxAcjdjaRbCGSpUk6Pryz7TzJEpF4/2kDis/ZRLKFSDYif0bkT5QpWUQkO5tIthDJEpH4dINNqCEhIlmUKdnZRHKFSI4yJT4J7ihT4hM4jpgkll8Rk9xsJrnCJKeaLd4RkXiEOgX86RCR3GwiuUIkdwYJT+ecBo/dUV7Fp28c2MwwOESk920+99vn7e57UguV3BklXnT2ziCxgGO9AynU4BCVpoktZHI2i+VDPEdkErsEiGW8J3UgjRocItM0sdWGBpQxZYPkfTaIWCVs8d7UITpNk1oI5Xz2q5Dqs1TeozrKsniP6kAyNThEqGliC6UcWmPKhobYgCI2IM8iTk0TW1jlYvas6FmJVSIIaIeE6AsAxAaHWDVNbOGVQxlUNmgx4nep/S36bOD5k0O0miTVF2L5LvuVo9WjHRC+AxHrQXo1eESsaWILtTxlUnySJxu0wKsnoPFuywOgDR5xa5rYwi7fzqbyZQ4DD5KsPht4APhF3PKFW55m+His+vamO0/U4rMnHlHLL6KWL9TymVpcqM0PX/jUNjHgm0tRg1/EK1945V1uUnxY7R1qUjT1x7sqD6b+Br+IWL7aikc5lfCqR2LBPr3eE7FE+19ELF+I5UP2LCeWp7k+PmzxYLtE70HKNfhFxPKFWD4iz8Z2wyJeiR4gooa1iFe+8MrTShTPVz3xSvSrRDi+oOvbvPKLeBUKrwLxiiMgEK94LxA64NcAUq8hLOJVKLwKCoRAIF7xmZWAeBWIV9yzYRGvQuFVoF0SHAQBZVsBrGL1gbKtyMUuolYo1AqUbfE+NtAcIB+7ZgPvBwKYGhzCImqFQq1AuRYfuQaLggDlWgHlWmERu0JhV0DsCg75FbErELv4zEBYxK5Q2BUo2xLbtBG7AmVbfJgdELvCInaFaks5sUv0Be05wYDIFRC5wiJyhUKuQLkWx0E2NIKA2CV6ArDtfAiL2BUKu0ICY4JsEH5t7krvA00YBi50EbliIVfsmqPXiPKsiLgVEbfiIm7Fwq1I3OI+jcQtvgIYiVtCrGqOB+IiasVCrUjU4nlWNshUOxLOeKxGopYQu4hasVArGiTWILFg41+fDULqImrFQq1I+yl4HxCJWnxIGGnqkFMrErU4COIiasVCrejaTcshqWBfYB/BwtYQFzErFmZFyrc4s7KBd1eRiCUaFpofjIuIFQuxYmh7lXItEatodjCCbYBDXESsWB2GImKJDotmB3liGNHsYETZVlxErFiIFYlYfEAYKdsSQUDZlji8RdQSp7cWUSsVaiVa0eJhkNp7LBJRizethGYH0yJqpUKtRNQSJ85QtpUN3KtJIamLqJUKtRKiViJqccAmDUIgEbXE+bhF1EqFWgnlWqk9Q5hohpBPECVwYHdIi6iVCrUSolayyK/tOcKEmJUWMSsVZqX2nouE1rQS2LLeZwMfCaRFxEqFWAnNECbKsjheEzGLTxEnlGWlRcxKhVkJrWmldpaVApIKNgkOaRGzUmFWikgqZVk8I0gRiQUbBYe0iFmpOspLzOLTbikhscQsEa+IWWkRs1RXH+slaonDXh2lW+K0V9cEWk/XZVpwsiwSXB3v7VDGRRaZc5NFopYs0sUnyyLJupKswTCGLHJJdrxHCNbtpzKcLEDwZvd587q5//5J7+q8b0fJlzj52VH2JY4mj/dIFxPLklAMYTZJcXXwtztDKDSiGBz97Yh/PAcji25EMSTaJL3VCeDOwYhAK19kEXTpySInkU+WJYqrw8Ad2q9BFulhcH64J4tYgBhOliV6q1PBXcgRwVlMFqkXTDv2ZGlFBETcJL3V4eCOiCXKFnQEOdkPo3Uwssg5upNlieLqlHCXsodlDCPSkUXEfU+WVgxD2E1RXJewULQiJtChIOtUh6J4rG8hFMM6FtMU17UsVPYxH0yosZoFH6aTRSQhPVka7IAFLaYprliniHWinxjLWvDFMTVWwpCKm1tABgXrWkzTW7FOoc2IZGmwThHrRBQryDpY4GKa4op1ilbJBJ1RmQtFrJMFTiDrYKWLaXor1inilmx1Dra6ZsrXj3cItYs4Vxe9wFUvFNqYWO6R8Uv3iEINsPbFNMUV6VSAEYzOdamxaIaMYGKdGBDDIhjTFFesUzQHKfuI9oYPNdbOkBFBpJO98CLS1fUwFJFORjBtRpQxARbberI0Rj+wKMYkxVVhDJXLWTQUayKd4IYmOooR8XiPiAlYHWOa4op0ua6FGK1p1fbWHVlkLSRwJGxQsETGNLUV5bRu++qT0pByY20NERFaQ/8u4lxVL0ONBTOkh9E+ELLI3RXjPULvIspVZTOUtjCCLdRrUZuje4TeRZSr6mcoDTM6jdbWyNLyL8zoYBGNaYor0mnf9tYnNZbSEP2w9jCGiXRiJAGraUxTXJFOU04nRsMakk43l+Z6NRbckB5exLmquIbSxDmpF20PIYvcH6I0JB2ssTFNcUU6TdQS8yi6vUeErrdiGGZ0sNTGJL1VuQ1liFli7GM65GED1un6co+IYVhzY5riinMGZnRUeUPopQlPERGoJMegYOmNaXor0uWiGaHhYQ09rNtR1KuxNofIj2ANjmmK62qFJisWbB4rcUjFIAvsySLrFS4iXVWOQxkLYxiSbqzhIWMYbPEfFKzJMU1xxToDWWfQ3kdl0EYSNRbtEOM1WJxjmuKKdca3I/ITWRrjYQNW83o1Vu8QWR2s0jFNccU6Q6wTdDZg/tIE2OpgTgfLdUzTW7HOoFNnaizaIauEgn3/PVkaeTOs2zFNccU6k2C7o6xOVIo1tP1fZB0G0g4W8JikuCrioWx7X6QCRTzUWPVDkMPCuUtYx2Oa2op0luYhRT88VvMQ/h0LgMiKrJQHypKsi1hXFfVQlrgl2txY2EMq1iiGLdhWOShY22Oa4op11kAfG6gYrtShqiCDgkU+pimuS/SifSdkaeRJ4z1SMW0+EeMJWO1jmuKKdpZoJxXTRhORdVCZEDkqtmDXyqBg4Y9piivaWZjZWZjZWZjZWSKk4DOsADJNcUU7i46rKVAHhK6LA649WVoeXkS7qhyIspB2Fq7WUQ0RoTfCGF7EuqoqiLLEOjHjakFeZ0HJxZ4sYnwJy4JMUluVBlEOnVxTY3EQMZYY64kIcjg4ewnrg0xTXLGOioTIftiBrG68Qxb1pqxOsBkWCvmu3uvqH8i8rT6v/746/fH+4nn9eHyz7urYKHbn/zfz/vth+/b+27E3+nV7OGxf6NXTevWw3p1eHWHwuN0e6MXp39qM/7fn9v9QSwMEFAAAAAgAAAAhAGIIViNcAQAAYgIAAA8AAAB4bC93b3JrYm9vay54bWyNUctOwzAQvPMVlu80iZVEtGpSiZeohIADtGcTbxqrjh3ZDmn/nnWqlCJx4GTPrmd2Z7xcHVpFvsA6aXRBk1lMCejKCKl3Bf14f7y+ocR5rgVXRkNBj+DoqrxaDsbuP43ZE+RrV9DG+24RRa5qoOVuZjrQ2KmNbblHaHeR6yxw4RoA36qIxXEetVxqelJY2P9omLqWFdybqm9B+5OIBcU9bu8a2TlaLmupYHMyRHjXvfAW1z4oShR3/kFID6KgKUIzwK+C7bvbXqoAsjijUXk2+WaJgJr3yr/japM6xsVSxvLwMrzaSBjcDylActhKLcxQUJZisscJJRklw3jfSuEbLOTx/Fx7ArlrfEHneR4H8ehCfcxvOokezbW+4tYRwT134PG/Qm+NPhI0tZB4sWuRBKVLlrTyLw674LBx+jSy4qrCJMIxirM0S+aU1L1Sd1h71c+Gj0MDaUqh/AZQSwMEFAAAAAgAAAAhAD/Y7yGxBQAAUxsAABMAAAB4bC90aGVtZS90aGVtZTEueG1s7VlNj9NGGL7zK0a+g+PEDtkVWbTJJtDCwmo3UHGc2BN7yNhjzUx2ya2CY6VKVWnVS6XeeqjaIoHUC/0121K1VOIv9PVHkvFmsmRhqxZBDoln/LzfH37HuXL1QczQIRGS8qRtOZdqFiKJzwOahG3rzqB/sWUhqXASYMYT0ramRFpXty5cwZsqIjFBQJ7ITdy2IqXSTduWPmxjeYmnJIF7Iy5irGApQjsQ+AjYxsyu12pNO8Y0sVCCY+B6ezSiPkGDjKW1NWPeY/CVKJlt+Ewc+LlEnSLHBmMn+5FT2WUCHWLWtkBOwI8G5IGyEMNSwY22Vcs/lr11xZ4TMbWCVqPr55+SriQIxvWcToTDOaHTdzcu78z51wv+y7her9ftOXN+OQD7PljqLGHdfsvpzHhqoOJymXe35tXcKl7j31jCb3Q6HW+jgm8s8O4SvlVrutv1Ct5d4L1l/Tvb3W6zgvcW+OYSvn95o+lW8TkoYjQZL6GzeM4jM4eMOLtuhLcA3polwAJla9lV0CdqVa7F+D4XfQDkwcWKJkhNUzLCPuC6OB4KijMBeJNg7U6x5culrUwWkr6gqWpbH6cYKmIBefX8x1fPn6JXz58cP3x2/PCX40ePjh/+bCC8jpNQJ3z5/Rd/f/sp+uvpdy8ff2XGSx3/+0+f/fbrl2ag0oEvvn7yx7MnL775/M8fHhvg2wIPdfiAxkSiW+QI7fMYbDMIIENxNopBhGmFAkeANAB7KqoAb00xM+E6pOq8uwIagAl4bXK/outBJCaKGoA3orgC3OWcdbgwmnMjk6WbM0lCs3Ax0XH7GB+aZHdPhLY3SSGTqYllNyIVNfcYRBuHJCEKZff4mBAD2T1KK37dpb7gko8UukdRB1OjSwZ0qMxE12kMcZmaFIRQV3yzexd1ODOx3yGHVSQUBGYmloRV3HgNTxSOjRrjmOnIm1hFJiUPpsKvOFwqiHRIGEe9gEhporktphV1b2DoRMaw77JpXEUKRccm5E3MuY7c4eNuhOPUqDNNIh37kRxDimK0x5VRCV6tkGwNccDJynDfpUSdrazv0DAyJ0h2ZyLKrl3pvzFNTmvGjEI3/tCMZ/BteDSZSuJkC16Fewcb7w6eJHsEcv1D3/3Qd9/HvruqltfttosGa+tzcc4vXjkkjyhjB2rKyE2Zt2YJSgd92MwXOdF8Jk8juCzFVXChwPk1Elx9QlV0EOEUxDi5hFCWrEOJUi7hJGCt5J0fJykYn+95szMgoLHa5UGx3dDPhnM2+SqUuqBGxmBdYY3LbyfMKYBrSnM8szTvVGm25k2oBoSzg7/TrBeiIWMwI0Hm94LBLCznHiIZ4YCUMXKMhjiNNd3Wer3XNGkbjbeTtk6QdHHuCnHeOUSpthQle7kcWVJdoSPQyqt7FvJx2rZGMEnBZZwCP5k1IMzCpG35qjTltcV80mBzWjq1lQZXRKRCqh0so4IqvzV7dZIs9K97buaH8zHA0I3W06LRcv5DLeyToSWjEfHVip3FsrzHJ4qIgyg4QkM2EfsY9HaL7AqohGdGfbYQUKFumXjVyi+r4OQrmrI6MEsjXPaklhb7Ap5fz3XIV5p69grd39CUxjma4r2/pmSZC2NrI8gPVDAGCIyyHG1bXKiIQxdKI+r3BQwOuSzQC0FZZCohlr1vznQlh4u+VfAomlwYqX0aIkGh06lIELKnSjtfw8yp68/XGaOyz8zVlWnxOySHhA2y6m1m9lsomnWT0hE57mTQbFN1DcP+/3jycVdMPqePBwtB7llmEVdr+tqjYOPtVDjjo7Zutrjurf2oTeHwgbIvaNxU+Gwx3w74PkQfzSdKBIl4sVWW33xzCDq3NOMyVv/uGLUIQWtFvM9z+NSc3Vjh7NPFvbmzPYOvvdNdbS+XqK0dZPLV0h9PfHgfZO/AQWnClCzeJj2Ao2Z39pcB8LEXpFsX/gFQSwMEFAAAAAgAAAAhAK/OEl3DAQAAUgQAAA0AAAB4bC9zdHlsZXMueG1svVTfi5wwEH7vXxHy3nMVetwV9SiFhUJbDm4LfY0aNTBJJBkXvb/+JkbdXWg56ENfzMw3M9/8cJL8adLAztJ5ZU3B07sDZ9LUtlGmK/iv0/HjA2cehWkEWCMLPkvPn8oPuccZ5EsvJTJiML7gPeLwOUl83Ust/J0dpCFLa50WSKrrEj84KRofgjQk2eFwn2ihDC/z1hr0rLajwYJnK1Dm/pWdBVBZKU/KvLZgHUOipzoWxAgto8dXAapyKoCt0ArmCGcBWCpa/bQy1gUwiRnit0r+R67l8BSkAG6bJaDMB4EonTmSwlb5NA+U3tDgI83i945358ScZp+uApaD8lbWNfSjt8wp36AyB9kiBTjV9eFEOyTBiGg1CY0SnTUCAuUWsQpEW0uAl7ANv9sb7qllZtRHjd+agtNahe43kQpaxUgTlcB/zRa5r2izf6JlU7vz/y06fT+aiWGA+Whjf1H7AqozWm4ti01lvXXqlVzDHtQEyLgIU7u2uHe39Hoztx1lYeUK/jPcILiqoxoVoDJ/mBlxNtNlXIsVRUUX9SYLcTSyFSPgaTcW/CL/kI0a9ePu9azOFlevi/w9LEt6v1RweQ3KN1BLAwQUAAAACAAAACEAbk2rXSUBAABQAgAAEQAAAGRvY1Byb3BzL2NvcmUueG1snZLLasMwEEX3/QqjvS3Z7iMY24G2ZNVAoSkt3Qlp4ohaDyS1jv++sp04CWRV0Ea6d87cGVQu97KNfsE6oVWF0oSgCBTTXKimQu+bVbxAkfNUcdpqBRXqwaFlfVMyUzBt4dVqA9YLcFEAKVcwU6Gd96bA2LEdSOqS4FBB3GorqQ9X22BD2TdtAGeE3GMJnnLqKR6AsZmJ6IDkbEaaH9uOAM4wtCBBeYfTJMUnrwcr3dWCUTlzSuF7A1etR3F2752YjV3XJV0+WkP+FH+uX97GUWOhhlUxQHXJWcEsUK9tXeLzS1hcS51fhxVvBfDHPuhX3g6DTHXAoxCgmOIelY/86XmzQnVGMhKT2zi925CHIl+E8zW0vKg/AeWhyb+JR8CU+/IT1H9QSwMEFAAAAAgAAAAhAM5Dm3KBAQAAOQMAABAAAABkb2NQcm9wcy9hcHAueG1snVPBahsxEL33KxbdY21MCcVoFYrTkENCDXaS81Q76xXRSkIzWex+fbVrvFm3hUJ1evPm8XgajdTtoXNFj4ls8JW4XpSiQG9Cbf2+Es+7+6svoiAGX4MLHitxRBK3+pPapBAxsUUqsoOnSrTMcSUlmRY7oEVu+9xpQuqAc5n2MjSNNXgXzHuHnuWyLG8kHhh9jfVVnAzFyXHV8/+a1sEM+ehld4zZT6uvMTprgPMl9ZM1KVBouPh2MOiUnDdVNtqieU+Wj7pUcl6qrQGH62ysG3CESn4Q6gFhmNkGbCKtel71aDikguzPPLWlKH4A4RCnEj0kC57FSXYqRuwicdKvIb1Ri8ik5ESOcK6dY/tZL0dBBpdCOQXJ+DLizrJD+t5sIPG/Eo8ZxCxjxwYSFTVw1vA854Rssn8VXET6LcQ6dBF8nrSc0KP1b/Qcd+EOGM9zvyTVtoWEdX6q6V0mQj3kCyQ36Nct+D3WZ82fjWFLXk4/QV8vF2U+43KcOSU/ll7/AlBLAQIAABQAAAAIAAAAIQCeDieWQgEAAI8EAAATAAAAAAAAAAEAAAAAAAAAAABbQ29udGVudF9UeXBlc10ueG1sUEsBAgAAFAAAAAgAAAAhAPKfSdrpAAAASwIAAAsAAAAAAAAAAQAAAAAAcwEAAF9yZWxzLy5yZWxzUEsBAgAAFAAAAAgAAAAhAJlwI8XfAAAAuQIAABoAAAAAAAAAAQAAAAAAhQIAAHhsL19yZWxzL3dvcmtib29rLnhtbC5yZWxzUEsBAgAAFAAAAAgAAAAhAGhFXh7yBwAA8ycAABgAAAAAAAAAAQAAAAAAnAMAAHhsL3dvcmtzaGVldHMvc2hlZXQxLnhtbFBLAQIAABQAAAAIAAAAIQB5Ur3rzw4AAPtnAAAYAAAAAAAAAAEAAAAAAMQLAAB4bC93b3Jrc2hlZXRzL3NoZWV0Mi54bWxQSwECAAAUAAAACAAAACEAYghWI1wBAABiAgAADwAAAAAAAAABAAAAAADJGgAAeGwvd29ya2Jvb2sueG1sUEsBAgAAFAAAAAgAAAAhAD/Y7yGxBQAAUxsAABMAAAAAAAAAAQAAAAAAUhwAAHhsL3RoZW1lL3RoZW1lMS54bWxQSwECAAAUAAAACAAAACEAr84SXcMBAABSBAAADQAAAAAAAAABAAAAAAA0IgAAeGwvc3R5bGVzLnhtbFBLAQIAABQAAAAIAAAAIQBuTatdJQEAAFACAAARAAAAAAAAAAEAAAAAACIkAABkb2NQcm9wcy9jb3JlLnhtbFBLAQIAABQAAAAIAAAAIQDOQ5tygQEAADkDAAAQAAAAAAAAAAEAAAAAAHYlAABkb2NQcm9wcy9hcHAueG1sUEsFBgAAAAAKAAoAhAIAACUnAAAAAA==" download="mtcars and iris datasets.xlsx">
#>   <button class="btn btn-warning"><i class="fa fa-save"></i> Download data</button>
#> </a>